pyproject.toml
[project]
name = "readingroomai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "moviepy>=2.1.2",
    "numpy>=2.2.3",
    "openai>=1.65.5",
    "pillow>=10.4.0",
]
